user table wip
