+++ frontend-nx/apps/edu-hub/components/Menu.tsx
@@ -79,6 +79,14 @@ export const Menu: FC<IProps> = ({ anchorElement, isVisible, setVisible }) => {
 
       {isAdmin && (
         <MenuItem onClick={closeMenu}>
+          <Link className="w-full text-lg" href="/manage/users-old">
+            {t('menu-administration-user')}
+          </Link>
+        </MenuItem>
+      )}
+
+      {isAdmin && (
+        <MenuItem onClick={closeMenu}>
           <Link className="w-full text-lg" href="/manage/courses">
             {t('menu-administration-courses')}
           </Link>
