+++ frontend-nx/apps/edu-hub/pages/manage/users/index.tsx
@@ -1,128 +1,29 @@
-import useTranslation from 'next-translate/useTranslation';
+// do not remove this https://github.com/nrwl/nx/issues/9017#issuecomment-1140066503
+import path from 'path';
+path.resolve('./next.config.js');
+
 import Head from 'next/head';
-import { FC, useCallback, useState } from 'react';
-import CommonPageHeader from '../../../components/common/CommonPageHeader';
-import EhMenuItem from '../../../components/common/EhMenuItem';
-import SearchBox from '../../../components/common/SearchBox';
+import { FC } from 'react';
 import { Page } from '../../../components/Page';
-import ManageUsersContent from '../../../components/ManageUsersContent';
 import { useIsAdmin, useIsLoggedIn } from '../../../hooks/authentication';
-import { StaticComponentProperty } from '../../../types/UIComponents';
 
-const Users: FC = () => {
-  const { t } = useTranslation('users');
-  const isLoggedIn = useIsLoggedIn();
+import ManageUsersContent from '../../../components/ManageUsersContentOld';
+
+const ManageUsers: FC = () => {
   const isAdmin = useIsAdmin();
+  const isLoggedIn = useIsLoggedIn();
+
   return (
     <>
-      <div className="max-w-screen-xl mx-auto">
-        <Head>
-          <title>EduHub | opencampus.sh</title>
-          <link rel="icon" href="/favicon.png" />
-        </Head>
-        <Page>
-          <div className="min-h-[77vh]">
-            {isLoggedIn && isAdmin && <UserDashboard t={t} />}
-          </div>
-        </Page>
-      </div>
+      <Head>
+        <title>EduHub | opencampus.sh</title>
+        <link rel="icon" href="/favicon.png" />
+      </Head>
+      <Page>
+        <div className="min-h-[77vh]">{isLoggedIn && isAdmin && <ManageUsersContent />}</div>
+      </Page>
     </>
   );
 };
 
-export default Users;
-
-interface IProps {
-  t: any;
-}
-const UserDashboard: FC<IProps> = ({ t }) => {
-  const menuItems: StaticComponentProperty[] = [
-    { key: -1, label: t('all'), selected: true },
-  ];
-  const [searchText, setSearchText] = useState('');
-
-  /* #region Callbacks */
-  const handleSearch = useCallback(
-    (text: string) => {
-      setSearchText(text);
-    },
-    [setSearchText]
-  );
-
-  const handleMenuItemClick = useCallback(
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    (property: StaticComponentProperty) => {},
-    []
-  );
-  /* #endregion */
-  return (
-    <>
-      <CommonPageHeader headline={t('headline')} />
-      <Menubar
-        t={t}
-        topMenuItems={menuItems}
-        handleSearch={handleSearch}
-        onMenuItemClick={handleMenuItemClick}
-        searchText={searchText}
-      />
-      <ManageUsersContent t={t} searchedText={searchText} />
-    </>
-  );
-};
-
-interface IMenubarProps {
-  t: any;
-  handleSearch: (text: string) => void;
-  searchText: string;
-  topMenuItems: StaticComponentProperty[];
-  onMenuItemClick?: (property: StaticComponentProperty) => void;
-}
-const Menubar: FC<IMenubarProps> = ({
-  t,
-  topMenuItems,
-  handleSearch,
-  searchText,
-  onMenuItemClick,
-}) => {
-  const [menuItems, setMenuItems] = useState(topMenuItems);
-
-  const updateMenuBar = useCallback(
-    (selected: StaticComponentProperty) => {
-      const newItems = menuItems.map((item) => {
-        if (selected.key === item.key) return { ...item, selected: true };
-        return { ...item, selected: false };
-      });
-      setMenuItems(newItems);
-    },
-    [menuItems, setMenuItems]
-  );
-
-  /* #region Callbacks */
-  const handleTabClick = useCallback(
-    (property: StaticComponentProperty) => {
-      updateMenuBar(property);
-      onMenuItemClick(property);
-    },
-    [updateMenuBar, onMenuItemClick]
-  );
-  /* #region */
-
-  return (
-    <div className="flex justify-between mb-5 text-white">
-      <div className="flex items-center space-x-5">
-        {menuItems.map((tab) => (
-          <EhMenuItem
-            key={tab.key}
-            property={tab}
-            onClickCallback={handleTabClick}
-          />
-        ))}
-      </div>
-      <SearchBox
-        onChangeCallback={handleSearch}
-        placeholder={t('userSearchPlaceHolder')}
-        searchText={searchText}
-      />
-    </div>
-  );
-};
+export default ManageUsers;
